medium/requirements.txt
langgraph
langchain
langchain-community
langchain-openai
python-dotenv
duckduckgo-search